requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
djangorestframework==3.7.7
idna==2.6
nltk==3.2.5
oauthlib==2.0.7
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
textblob==0.15.1
urllib3==1.22
